requirements.txt
PyMuPDF==1.24.5
pdfminer.six==20221105
sentence-transformers==2.7.0
scikit-learn==1.5.0
numpy==1.26.4
summa==1.0.0
pytesseract
PyPDF2